fix: restore wgpu sprite rendering and fix all SDK example builds
The 0.0.833 wgpu migration broke sprite rendering because naga 28's GLSL
frontend rejects bare `uniform` declarations without `layout(binding=X)`.
This caused shader creation to fail silently — no sprites rendered, only
the clear color was visible.

Fix: add native WGSL sprite shaders with proper @group/@binding annotations,
naga WGSL reflection for uniform struct layout extraction, and per-draw
dynamic uniform buffer offsets so each sprite gets its own transform.

Also fixes all SDK example builds across 9 languages:
- C# (7 examples): switch from stale NuGet 0.0.826 to ProjectReference
- Swift (3 examples): fix codegen for bytes/string length types, skip
  methods with complex FFI pointer patterns, add out_buffer metadata
- Kotlin (3 examples): fix Gradle fileTree dependency, add lifecycle
  methods (beginFrame/endFrame/deltaTime) to codegen and JNI layer
- Go (3 examples): enable rapier2d/rapier3d in default Cargo features
- C/C++ smoke tests: add missing CMakeLists.txt build files
- Swift codegen: use FFI manifest for correct integer casts

Relates to #114 (Phase 3 gate: wgpu works), #475 (Phase 0: render perf)

+++ examples/csharp/goud_jumper/GameManager.cs
@@ -149,7 +149,7 @@ public class GameManager
             spriteHeight,
             0f,
             sourceRect.X, sourceRect.Y, sourceRect.Width, sourceRect.Height,
-            srcMode: SrcRectMode.Normalized);
+            srcMode: (uint)SrcRectMode.Normalized);
     }
 
     private void UpdatePlayerPosition(float deltaTime)
